requirements.txt
attrs==22.1.0
contextlib2==21.6.0
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
schema==0.7.5
tomli==2.0.1